web/backend/requirements.txt
flask
flask-cors
paho-mqtt
stable-baselines3
sqlite3
matplotlib
numpy
torch